+++ defeatbeta_api/data/sql/select_news_list_by_symbol.sql
@@ -1,0 +1,4 @@
+SELECT uuid, related_symbols, title, publisher, report_date, type, link
+FROM '{url}'
+WHERE related_symbols = '{ticker}'
+ORDER BY report_date ASC